refact: move ark utils to single file
