+++ apps/mobile/hooks/useArkSendNavigation.ts
@@ -3,10 +3,8 @@ import { useRouter } from 'expo-router'
 import { toast } from 'sonner-native'
 
 import { t } from '@/locales'
-import {
-  type ArkDestinationDraft,
-  parseArkDestination
-} from '@/utils/arkDestination'
+import { type ArkDestinationDraft } from '@/types/models/Ark'
+import { parseArkDestination } from '@/utils/ark'
 import { type DetectedContent } from '@/utils/contentDetector'
 import { getLNURLType } from '@/utils/lnurl'
